dsashudaskasd

+++ src/views/CheckoutView.vue
@@ -1,10 +1,11 @@
 <script setup>
-import { onMounted } from 'vue'
+import { onMounted, ref, defineProps} from 'vue'
 import { useRouter } from 'vue-router'
 import Slider from '@/components/Slider.vue'
 import NationalityDropdown from '@/components/NationalityDropdown.vue'
 import CheckoutHelper from '@/utilities/CheckoutHelper'
 import DashboardHelper from '@/utilities/DashboardHelper'
+import AlertCard from '@/components/AlertCard.vue'
 const route = useRouter()
 
 const {
@@ -18,33 +19,104 @@ const {
   reduceTicket,
   selectedDate,
   discountValue,
+  cashbackValue,
   biayaLayanan,
   biayaJasa,
   formatCurrency,
-  formattedTotalHarga,
-  formattedTotalTagihan,
+  totalHarga,
+  totalTagihan,
+  totalBiaya,
   totalTicketCount,
-  createTransaction
+  createTransaction,
+  selectedNationality,
+  
 } = CheckoutHelper
 
 const { checkSessionStorage, isMancanegara } = DashboardHelper
 
-const checkoutTransaction = () => {
-  createTransaction()
-  sessionStorage.clear()
-  setTimeout(() => {
-    route.push('/')
-  }, 3000)
+const showAlert = ref(false)
+const alertType = ref('')
+const alertTitle = ref('')
+const alertMessage = ref('')
+
+
+
+const checkoutTransaction = async () => {
+  const invalid = checkValidTransaction()
+  console.log(totalTicketCount.value)
+  if(totalTicketCount.value < 1){
+    console.log('Nyobain ererror')
+    showAlert.value = true
+    alertTitle.value = 'Error'
+    alertType.value = 'danger' // Set your alert type
+    alertMessage.value = `Pilih tipe tiket terlebih dahulu`
+
+    setTimeout(() => {
+    showAlert.value = false
+    }, 1200)
+    return
+  }
+  if (invalid.length > 0) {
+    showAlert.value = true
+    alertTitle.value = 'Error'
+    alertType.value = 'danger' // Set your alert type
+    alertMessage.value = `Isi kolom ${invalid.join(', ')} terlebih dahulu.` // Set your alert message
+    
+    setTimeout(() => {
+    showAlert.value = false
+    }, 1200)
+    return
+  }
+  
+  try {
+    await createTransaction()
+    sessionStorage.clear()
+    setTimeout(() => {
+      route.push('/')
+    }, 3000)
+  } catch (error) {
+    console.error('Gagal melakukan transaksi:', error)
+    // Tampilkan pesan kesalahan atau lakukan tindakan yang sesuai jika transaksi gagal
+  }
+}
+
+
+const checkValidTransaction = () => {
+  const invalid = []
+
+  if (isMancanegara.value) {
+    if (!selectedNationality.value) {
+      invalid.push('Kewarganegaraan')
+    }
+  }
+  if (!selectedDate.value) {
+    invalid.push(' Tanggal Pemesanan')
+  }
+
+  if (!paymentSelection.value) {
+    invalid.push(' Metode Pembayaran')
+  }
+
+  // Pastikan semua input telah diisi sesuai dengan kondisi
+  return invalid
 }
 
 onMounted(() => {
   getItemsFromSessionStorage()
   checkSessionStorage()
 })
+
+
 </script>
 
 <template>
   <main>
+    <AlertCard
+      :showAlert="showAlert"
+      :alertTitle="alertTitle"
+      :alertType="alertType"
+      :alertMessage="alertMessage"
+    />
     <div class="checkout__container sm-sd-2">
       <div class="checkout__form-container">
         <div class="order-details__container">
@@ -84,7 +156,7 @@ onMounted(() => {
                   <div class="order-details__ticket" v-if="item.amount > 0">
                     <div class="order-details__ticket-items">
                       <p>{{ item.name }} ({{ item.category }})</p>
-                      <span>Rp {{ formatCurrency(item.price) }},00</span>
+                      <span>{{ formatCurrency(item.price) }},00</span>
                     </div>
                     <div class="order-details__ticket-value">
                       <button @click="reduceTicket(index)" type="button">
@@ -101,7 +173,11 @@ onMounted(() => {
               <div class="pricings-slider__container">
                 <div class="discount__slider">
                   <p class="fw-500">Diskon</p>
-                  <Slider v-model:discountValue="discountValue" />
+                  <Slider v-model:targetValue="discountValue" />
+                </div>
+                <div class="discount__slider">
+                  <p class="fw-500">Cashback</p>
+                  <Slider v-model:targetValue="cashbackValue" />
                 </div>
               </div>
               <div class="order-details__content">
@@ -158,40 +234,62 @@ onMounted(() => {
             <p class="fs-h5">Ringkasan Booking</p>
             <div class="checkout__details-pricing-container">
               <p class="fw-700 fs-h6">Total Pemesanan</p>
-              <div
-                v-if="items.length > 1"
-                v-for="(item, index) in items"
-                :key="index"
-              >
+              <div v-if="items.length > 1" v-for="(item, index) in items" :key="index">
                 <div class="checkout__details-pricing" v-if="item.amount > 0">
                   <p>{{ item.name }} ({{ item.category }}) x {{ item.amount }}</p>
-                  <p>Rp {{ (item.price * item.amount).toLocaleString('id-ID') }},00</p>
+                  <p>{{ formatCurrency(item.price * item.amount) }}</p>
                 </div>
-              </div>
-              <div class="checkout__details-pricing" v-if="discountValue > 0">
-                <p>Diskon</p>
-                <p>{{ discountValue }}%</p>
               </div>
               <div class="checkout__details-pricing">
                 <p v-if="items.length > 1">Total Tiket ({{ totalTicketCount }} Tiket)</p>
                 <p v-else>Jumlah Tiket ({{ totalTicketCount }} Tiket)</p>
-                <p>{{ formattedTotalHarga }}</p>
+                <p>{{ formatCurrency(totalHarga) }}</p>
               </div>
             </div>
             <div class="checkout__details-pricing-container">
               <p class="fw-700 fs-h6">Biaya Transaksi</p>
               <div class="checkout__details-pricing">
                 <p>Biaya Layanan</p>
-                <p>Rp {{ formatCurrency(biayaLayanan) }}</p>
+                <p>{{ formatCurrency(biayaLayanan) }}</p>
               </div>
               <div class="checkout__details-pricing">
                 <p>Biaya Jasa Aplikasi</p>
-                <p>Rp {{ formatCurrency(biayaJasa) }}</p>
+                <p>{{ formatCurrency(biayaJasa) }}</p>
               </div>
             </div>
-            <div class="checkout__details-total">
+            <div
+              class="checkout__details-pricing-container"
+              v-if="discountValue > 0 || cashbackValue > 0"
+            >
+              <p class="fw-700 fs-h6">Potongan Harga</p>
+              <div class="checkout__details-pricing" v-if="discountValue > 0">
+                <p>Diskon</p>
+                <p>
+                  {{ formatCurrency((totalHarga * discountValue) / 100) }} ({{ discountValue }}%)
+                </p>
+              </div>
+              <div class="checkout__details-pricing" v-if="cashbackValue > 0">
+                <p>Cashback</p>
+                <p>
+                  {{ formatCurrency((totalHarga * cashbackValue) / 100) }} ({{ cashbackValue }})%
+                </p>
+              </div>
+            </div>
+            <div class="checkout__details-total flex fd-row align-items-center justify-content-sb">
               <p class="fw-700 fs-h5">Total Tagihan</p>
-              <p class="fw-700 fs-h6">{{ formattedTotalTagihan }}</p>
+              <div class="checkout__details-total--final flex fd-col align-items-f-end">
+                <p
+                  class="fw-700 fs-h6"
+                  :class="{
+                    'checkout__details-total--strikethrough': discountValue > 0 || cashbackValue > 0
+                  }"
+                >
+                  {{ formatCurrency(totalBiaya) }}
+                </p>
+                <p class="fw-700 fs-h6" v-if="discountValue > 0 || cashbackValue > 0">
+                  {{ formatCurrency(totalTagihan) }}
+                </p>
+              </div>
             </div>
           </form>
         </div>
@@ -455,9 +553,6 @@ main {
 }
 
 .checkout__details-total {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
   border-top: solid 0.5px #ccc;
   padding: 0.5rem 0;
 }
@@ -488,5 +583,12 @@ main {
 
 .checkout__btn-order i {
   font-size: 20px;
+}
+
+.checkout__details-total--strikethrough {
+  font-size: 16px;
+  line-height: 24px;
+  font-weight: 500;
+  text-decoration: line-through;
 }
 </style>
